Uses geonorm from Maven Central.
