+++ src/main/resources/eidos.conf
@@ -63,9 +63,6 @@ ontologies {
 }
 
 geonorm {
-  taggerModelPath       = /org/clulab/geonorm/model/geonorm_model.pb
-  taggerWordToIndexPath = /org/clulab/wm/eidos/${EidosSystem.language}/context/word2idx_file.txt
-  normalizerModelPath   = /org/clulab/wm/eidos/${EidosSystem.language}/context/geonames-reranker.model
   geoNamesIndexURL      = "http://clulab.cs.arizona.edu/models/geonames-index.zip"
   geoNamesIndexPath     = ${EidosSystem.cacheDir}/geonames/index
 }
